requirements.txt
python-dotenv==1.1.1 
pydantic==2.11.7
langchain==0.3.27
langchain-community==0.3.27
openai==1.98.0 
langchain-openai==0.3.28 
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
langchain-huggingface==0.3.1 
